requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
audioread==3.1.0
av==14.0.1
blis==1.3.3
braceexpand==0.1.7
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorlog==6.9.0
confection==0.1.5
cymem==2.0.13
decorator==5.2.1
docopt==0.6.2
dora_search==0.1.12
einops==0.8.0
exceptiongroup==1.3.1
filelock==3.20.0
flashy==0.0.2
fsspec==2025.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h5py==3.15.1
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
laion_clap==1.1.6
langcodes==3.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
networkx==3.4.2
nnAudio==0.3.3
numba==0.64.0
numpy==1.23.5
omegaconf==2.3.0
open_clip_torch==2.29.0
packaging==24.2
pandas==2.3.3
platformdirs==4.9.2
pooch==1.9.0
preshed==3.0.12
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
retrying==1.4.2
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.2
soundfile==0.13.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
submitit==1.5.4
sympy==1.13.1
thinc==8.3.10
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.15.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==4.38.2
treetable==0.2.6
triton==3.1.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
webdataset==1.0.2
wget==3.2
wrapt==2.1.1
